filepath: do not run TestEvalSymlinks on Windows
R=golang-dev, rsc1
CC=adg, golang-dev
https://golang.org/cl/4274069

+++ src/pkg/path/filepath/path_test.go
@@ -440,6 +440,10 @@ var EvalSymlinksTests = []EvalSymlinksTest{
 }
 
 func TestEvalSymlinks(t *testing.T) {
+	// Symlinks are not supported under windows.
+	if runtime.GOOS == "windows" {
+		return
+	}
 	defer os.RemoveAll("test")
 	for _, d := range EvalSymlinksTestDirs {
 		var err os.Error
